Internal: Moving things around, small refactoring [ED-19765]
